+++ LocalScrollWidget/LocalScrollWidgetBundle.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+import WidgetKit
+
+@main
+struct LocalScrollWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        ProcessingLiveActivityWidget()
+    }
+}